+++ __MOCKS__/Table/header.js
@@ -1,0 +1,20 @@
+ const HEADERS_TABLE = {
+  name: {
+    value: "Nome",
+    width: "28%"
+  },
+  notProcessed: {
+    value: "Não processado",
+    width: "18%"
+  },
+  paymentDate: {
+    value: "Data pgto",
+    width: "18%"
+  },
+  status: {
+    value: "Status",
+    width: "34%"
+  }
+};
+
+export default HEADERS_TABLE;