+++ __MOCKS__/Table/helpers/index.js
@@ -1,0 +1,2 @@
+export { default as sorts } from "./sorts";
+export { default as parseLotsToDataTable } from "./parseLots";