+++ __MOCKS__/Table/index.js
@@ -1,0 +1,2 @@
+export { default as HEADERS_TABLE } from "./header";
+export { default as tableMock } from "./tableTransfersMockData";