+++ src/components/Button/index.test.js
@@ -1,16 +1,16 @@
 import React from "react";
-import { shallow, render, mount } from "enzyme";
+import { shallow, render } from "enzyme";
 import Button from "./index";
 import theme from "../../styles/theme";
 
 describe("Button tests", () => {
   it("should render the main button", () => {
-    const wrapper = shallow(<Button>Render</Button>);
+    const wrapper = shallow(<Button onClick={() => {}}>Render</Button>);
     expect(wrapper).toExist();
   });
 
   it("should render children", () => {
-    const wrapper = render(<Button>Render</Button>);
+    const wrapper = render(<Button onClick={() => {}}>Render</Button>);
     expect(wrapper.html()).toBe("Render");
   });